refactor: 베팅 패키지 이동 반영
